+++ app/src/main/java/com/example/todo/view/OnChangeTaskStatusListener.kt
@@ -3,5 +3,5 @@ package com.example.todo.view
 import android.view.View
 
 interface OnChangeTaskStatusListener {
-    fun onChangeTaskStatus(id : Int, status : Boolean)
+    fun onChangeTaskStatus(id : Long, status : Boolean)
 }
