+++ app/src/main/java/com/example/todo/view/OnEditClickListener.kt
@@ -1,5 +1,5 @@
 package com.example.todo.view
 
 interface OnEditClickListener {
-    fun onClick(id : Int);
+    fun onClick(id : Long);
 }
